moved notification into rightmesh controller
